migrate .vscode-test.js to ES module
