+++ .vscode/launch.json
@@ -20,7 +20,7 @@
       "name": "Integration Tests",
       "type": "extensionHost",
       "request": "launch",
-      "testConfiguration": "${workspaceFolder}/.vscode-test.js",
+      "testConfiguration": "${workspaceFolder}/.vscode-test.mjs",
       "testConfigurationLabel": "integrationTests",
       "args": ["--profile=testing-debug"],
       "outFiles": ["${workspaceFolder}/dist/**/*.js"],
@@ -33,7 +33,7 @@
       "name": "Code Workspace Tests",
       "type": "extensionHost",
       "request": "launch",
-      "testConfiguration": "${workspaceFolder}/.vscode-test.js",
+      "testConfiguration": "${workspaceFolder}/.vscode-test.mjs",
       "testConfigurationLabel": "codeWorkspaceTests",
       "args": ["--profile=testing-debug"],
       "outFiles": ["${workspaceFolder}/dist/**/*.js"],
@@ -46,7 +46,7 @@
       "name": "Unit Tests",
       "type": "extensionHost",
       "request": "launch",
-      "testConfiguration": "${workspaceFolder}/.vscode-test.js",
+      "testConfiguration": "${workspaceFolder}/.vscode-test.mjs",
       "testConfigurationLabel": "unitTests",
       "outFiles": ["${workspaceFolder}/dist/**/*.js"],
       "preLaunchTask": "compile-tests"
